Few more rules added (filenaming)
- extracted vk api core to separate interface
