+++ src/server/vk/index.ts
@@ -1,0 +1,1 @@
+export { vkTool } from "./vk-tool";
